+++ src/lib/layout/search/utils.ts
@@ -39,3 +39,28 @@ export const getRouteOptions = (
       return null;
   }
 };
+
+export const getNextCursor = (
+  key: string,
+  current: Option<number>,
+  lastIndex: number
+) => {
+  switch (key) {
+    case "ArrowUp":
+      if (current === undefined) return lastIndex;
+      return current <= 0 ? lastIndex : current - 1;
+    case "ArrowDown":
+      if (current === undefined) return 0;
+      return current >= lastIndex ? 0 : current + 1;
+    default:
+      return undefined;
+  }
+};
+
+export const generateQueryObject = (
+  params: string[],
+  value: string | string[]
+) =>
+  typeof value === "string"
+    ? { [params[0]]: value }
+    : params.reduce((acc, curr, idx) => ({ ...acc, [curr]: value[idx] }), {});